mslearn-ai-agents/Labfiles/02-build-ai-agent/Python/requirements.txt
python-dotenv 
azure-identity